Fix(auth): Downgrade React to v18 to resolve compatibility issues
The application was experiencing issues with user registration and
authentication due to an incompatibility between NextAuth.js v4 and React 19.

This commit resolves the issue by downgrading React and React-DOM to
version 18.2.0, which is a stable version compatible with the rest of
the technology stack.

Additionally, this commit fixes the security middleware by activating the
correct implementation that enforces role-based access control for
protected routes.
