+++ package.json
@@ -14,17 +14,17 @@
       "chart.js": "^4.4.0",
       "next": "15.5.2",
       "next-auth": "^4.24.11",
-      "react": "19.1.0",
+      "react": "^18.2.0",
       "react-chartjs-2": "^5.2.0",
-      "react-dom": "19.1.0"
+      "react-dom": "^18.2.0"
    },
    "devDependencies": {
       "@eslint/eslintrc": "^3",
       "@tailwindcss/postcss": "^4",
       "@types/bcryptjs": "^2.4.6",
       "@types/node": "^20",
-      "@types/react": "^19",
-      "@types/react-dom": "^19",
+      "@types/react": "^18.2.17",
+      "@types/react-dom": "^18.2.7",
       "eslint": "^9",
       "eslint-config-next": "15.5.2",
       "prisma": "^5.22.0",
