--- src/middleware1.ts
@@ -1,23 +1,0 @@
-import { withAuth } from "next-auth/middleware";
-
-export default withAuth({
-  callbacks: {
-    authorized: ({ token, req }) => {
-      // Exemple: restreindre /admin aux ADMIN
-      const path = req.nextUrl.pathname;
-      if (path.startsWith("/admin")) {
-        return token?.role === "ADMIN";
-      }
-      return !!token;
-    },
-  },
-});
-
-export const config = {
-  matcher: [
-    "/profile/:path*",
-    "/reports/:path*",
-    "/players/:path*",
-    "/admin/:path*",
-  ],
-};
